refactor: move analyze_sarif_fields out of cli to fix inverted dep (#95)
core/phases/inspect_phase.py imported from cli/inspect/sarif_fields, creating a
dependency inversion (core -> cli). Move all functions to
utils/sarif_field_analysis.py, keep cli/inspect/sarif_fields.py as a deprecation
shim, and update all consumers (inspect_phase.py, cli/inspect/__init__.py).

Add tests/unit/utils/test_sarif_field_analysis.py (import-path + functional),
tests/unit/test_no_core_to_cli_imports.py (AST scan locks in the fix), and
update existing inspect-phase test patches to the new canonical location.
